pyproject.toml
[tool.poetry]
name = "dominions-slack-bot"
version = "0.1.0"
description = "A slack bot to help manage and update dominions 5 games"
authors = ["Steven Kessler <stvnksslr@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
slack-bolt = "^1.11.4"
tortoise-orm = { extras = ["accel"], version = "^0.18.1" }
aiohttp = "^3.8.1"
uvloop = "^0.16.0"
loguru = "^0.6.0"
pydantic = "^1.9.0"
uvicorn = "^0.17.4"
python-dotenv = "^0.19.2"


[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
black = "^21.12b0"
pytest-cov = "^2.8"
pytest-sugar = "^0.9.2"
pep8-naming = "^0.9.0"
flake8 = "^3.7"
flake8-pytest = "^1.3"
flake8-bandit = "^2.1.2"
flake8-functions = "^0.0.2"
flake8-variables-names = "^0.0.3"
flake8-bugbear = "^19.8.0"
flake8-eradicate = "^0.2.3"
pytest-asyncio = "^0.14.0"
flake8-comprehensions = "^3.1.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
